repos editable by permitted users.

+++ wsgi/myproject/issueview/views.py
@@ -303,16 +303,21 @@ def issues_authorize(request):
  
 
 @login_required(login_url=('/login/'))
-def issues_repo_delete(request, board, repoid):
-  repo = Repository.objects.filter(board__user=request.user).filter(pk=repoid)
+def issues_repo_delete(request, boardid, repoid):
+  board = Board.objects.filter(pk=boardid)
+  if board.user != request.user and len(ReadPermissions.objects.filter(username=request.user.username).filter(board_id = boardid)) == 0:
+    ret = HttpResponseRedirect('/issueview/board/show/')
+    add_never_cache_headers(ret)
+    return ret 
+  repo = Repository.objects.filter(board__user=board.user).filter(pk=repoid)
   if len(repo) == 0:
-    ret = HttpResponseRedirect('/issueview/repos/'+board+"/")
+    ret = HttpResponseRedirect('/issueview/repos/'+boardid+"/")
     add_never_cache_headers(ret)
     return ret 
   repo[0].delete()
   filtstring = request.GET.get("filter","")
   if len(filtstring):
-    ret = HttpResponseRedirect('/issueview/repos/'+board+"/"+"?filter="+filtstring)
+    ret = HttpResponseRedirect('/issueview/repos/'+boardid+"/"+"?filter="+filtstring)
   else:
     ret = HttpResponseRedirect('/issueview/repos/'+board+"/")
   add_never_cache_headers(ret)
